+++ packages/features/insights/server/__tests__/events.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "vitest";
 
 import dayjs from "@calcom/dayjs";
 
-import { EventsInsights } from "../events";
+import { getTimeView, getDateRanges, formatPeriod } from "../insightsDateUtils";
 
 describe("EventsInsights", () => {
   describe("getDateRanges", () => {
@@ -13,7 +13,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-01T00:00:00.000Z"; // Beginning of May 1st UTC
         const endDate = "2025-05-03T23:59:59.999Z"; // End of May 3rd UTC
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -50,7 +50,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-01T00:00:00.000Z"; // Thursday
         const endDate = "2025-05-25T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -99,7 +99,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-15T00:00:00.000Z";
         const endDate = "2025-07-15T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -136,7 +136,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-06-15T00:00:00.000Z";
         const endDate = "2027-03-15T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -173,7 +173,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-01T00:00:00.000Z";
         const endDate = "2025-05-01T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -202,7 +202,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-03-29T23:00:00.000Z"; // March 30th 00:00 Paris time
         const endDate = "2025-03-31T21:59:59.999Z"; // March 31st 23:59:59 Paris time
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -237,7 +237,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-15T22:00:00.000Z"; // May 16th 00:00 Paris time
         const endDate = "2025-05-29T21:59:59.999Z"; // May 29th 23:59:59 Paris time
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -274,7 +274,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-31T22:00:00.000Z"; // June 1st 00:00 Paris time
         const endDate = "2025-07-31T21:59:59.999Z"; // July 31st 23:59:59 Paris time
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -309,7 +309,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-14T15:00:00.000Z"; // May 15th 00:00 Seoul time
         const endDate = "2025-05-16T14:59:59.999Z"; // May 16th 23:59:59 Seoul time
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -340,7 +340,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-11T15:00:00.000Z"; // May 12th 00:00 Seoul time (Monday)
         const endDate = "2025-05-25T14:59:59.999Z"; // May 25th 23:59:59 Seoul time (Sunday)
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -377,7 +377,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-04-30T15:00:00.000Z"; // May 1st 00:00 Seoul time
         const endDate = "2025-06-30T14:59:59.999Z"; // June 30th 23:59:59 Seoul time
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -410,7 +410,7 @@ describe("EventsInsights", () => {
       const endDate = "2025-05-03T23:59:59.999Z";
       const timeZone = "UTC";
 
-      const ranges = EventsInsights.getDateRanges({
+      const ranges = getDateRanges({
         startDate,
         endDate,
         timeZone,
@@ -433,7 +433,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-01T00:00:00.000Z"; // Thursday
         const endDate = "2025-05-14T23:59:59.999Z"; // Wednesday
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -470,7 +470,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-01T00:00:00.000Z"; // Thursday
         const endDate = "2025-05-14T23:59:59.999Z"; // Wednesday
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -507,7 +507,7 @@ describe("EventsInsights", () => {
         const startDate = "2025-05-01T00:00:00.000Z"; // Thursday
         const endDate = "2025-05-14T23:59:59.999Z"; // Wednesday
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -546,7 +546,7 @@ describe("EventsInsights", () => {
     describe("Day View", () => {
       describe("Beginning of data (wholeStart === start)", () => {
         it("should always show month for the first day when same year", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-15"),
             end: dayjs("2024-01-15"),
             timeView: "day",
@@ -557,7 +557,7 @@ describe("EventsInsights", () => {
         });
 
         it("should always show month for the first day when different years", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-15"),
             end: dayjs("2024-01-15"),
             timeView: "day",
@@ -570,7 +570,7 @@ describe("EventsInsights", () => {
 
       describe("First day of month (start.date() === 1)", () => {
         it("should show month for 1st day of month when same year", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-02-01"),
             end: dayjs("2024-02-01"),
             timeView: "day",
@@ -581,7 +581,7 @@ describe("EventsInsights", () => {
         });
 
         it("should show month for 1st day of month when different years", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-02-01"),
             end: dayjs("2024-02-01"),
             timeView: "day",
@@ -594,7 +594,7 @@ describe("EventsInsights", () => {
 
       describe("Regular days (not first day, not 1st of month)", () => {
         it("should omit month for regular days when same year", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-16"),
             end: dayjs("2024-01-16"),
             timeView: "day",
@@ -605,7 +605,7 @@ describe("EventsInsights", () => {
         });
 
         it("should omit month but show year for regular days when different years", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-16"),
             end: dayjs("2024-01-16"),
             timeView: "day",
@@ -618,7 +618,7 @@ describe("EventsInsights", () => {
 
       describe("Edge cases", () => {
         it("should show month when start is both beginning of data AND 1st of month", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-01"),
             end: dayjs("2024-01-01"),
             timeView: "day",
@@ -630,7 +630,7 @@ describe("EventsInsights", () => {
 
         it("should show month for last day of month if it's 1st of month (edge case)", () => {
           // This tests the case where start.date() === 1 takes precedence
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-03-01"),
             end: dayjs("2024-03-01"),
             timeView: "day",
@@ -641,7 +641,7 @@ describe("EventsInsights", () => {
         });
 
         it("should omit month for end of month when not 1st and not beginning", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-31"),
             end: dayjs("2024-01-31"),
             timeView: "day",
@@ -662,7 +662,7 @@ describe("EventsInsights", () => {
           for (let i = 0; i <= 5; i++) {
             const currentDay = wholeStart.add(i, "day");
             results.push(
-              EventsInsights.formatPeriod({
+              formatPeriod({
                 start: currentDay,
                 end: currentDay,
                 timeView: "day",
@@ -684,7 +684,7 @@ describe("EventsInsights", () => {
           let currentDay = wholeStart;
           while (currentDay.isBefore(wholeEnd) || currentDay.isSame(wholeEnd)) {
             results.push(
-              EventsInsights.formatPeriod({
+              formatPeriod({
                 start: currentDay,
                 end: currentDay,
                 timeView: "day",
@@ -707,7 +707,7 @@ describe("EventsInsights", () => {
           let currentDay = wholeStart;
           while (currentDay.isBefore(wholeEnd) || currentDay.isSame(wholeEnd)) {
             results.push(
-              EventsInsights.formatPeriod({
+              formatPeriod({
                 start: currentDay,
                 end: currentDay,
                 timeView: "day",
@@ -722,7 +722,7 @@ describe("EventsInsights", () => {
         });
 
         it("should format a single day range correctly", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-15"),
             end: dayjs("2024-01-15"),
             timeView: "day",
@@ -738,7 +738,7 @@ describe("EventsInsights", () => {
     describe("Week View", () => {
       describe("Same month", () => {
         it("should format dates without year when wholeStart and wholeEnd are same year", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-01"),
             end: dayjs("2024-01-07"),
             timeView: "week",
@@ -749,7 +749,7 @@ describe("EventsInsights", () => {
         });
 
         it("should format dates with year when wholeStart and wholeEnd are different years", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-01"),
             end: dayjs("2024-01-07"),
             timeView: "week",
@@ -762,7 +762,7 @@ describe("EventsInsights", () => {
 
       describe("Different months", () => {
         it("should format dates without year when wholeStart and wholeEnd are same year", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-29"),
             end: dayjs("2024-02-04"),
             timeView: "week",
@@ -773,7 +773,7 @@ describe("EventsInsights", () => {
         });
 
         it("should format dates with year when wholeStart and wholeEnd are different years", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2024-01-29"),
             end: dayjs("2024-02-04"),
             timeView: "week",
@@ -786,7 +786,7 @@ describe("EventsInsights", () => {
 
       describe("Different years", () => {
         it("should format dates with respective years when start and end span different years", () => {
-          const result = EventsInsights.formatPeriod({
+          const result = formatPeriod({
             start: dayjs("2023-12-31"),
             end: dayjs("2024-01-06"),
             timeView: "week",
@@ -860,27 +860,27 @@ describe("EventsInsights", () => {
 
   describe("getTimeView", () => {
     it("should return year for ranges over 365 days", () => {
-      const result = EventsInsights.getTimeView("2024-01-01T00:00:00.000Z", "2025-02-01T00:00:00.000Z");
+      const result = getTimeView("2024-01-01T00:00:00.000Z", "2025-02-01T00:00:00.000Z");
       expect(result).toBe("year");
     });
 
     it("should return month for ranges between 90 and 365 days", () => {
-      const result = EventsInsights.getTimeView("2024-01-01T00:00:00.000Z", "2024-05-01T00:00:00.000Z");
+      const result = getTimeView("2024-01-01T00:00:00.000Z", "2024-05-01T00:00:00.000Z");
       expect(result).toBe("month");
     });
 
     it("should return week for ranges between 14 and 90 days", () => {
-      const result = EventsInsights.getTimeView("2024-01-01T00:00:00.000Z", "2024-02-01T00:00:00.000Z");
+      const result = getTimeView("2024-01-01T00:00:00.000Z", "2024-02-01T00:00:00.000Z");
       expect(result).toBe("week");
     });
 
     it("should return day for ranges under 14 days", () => {
-      const result = EventsInsights.getTimeView("2024-01-01T00:00:00.000Z", "2024-01-10T00:00:00.000Z");
+      const result = getTimeView("2024-01-01T00:00:00.000Z", "2024-01-10T00:00:00.000Z");
       expect(result).toBe("day");
     });
 
     it("should handle same day range", () => {
-      const result = EventsInsights.getTimeView("2024-01-01T00:00:00.000Z", "2024-01-01T23:59:59.999Z");
+      const result = getTimeView("2024-01-01T00:00:00.000Z", "2024-01-01T23:59:59.999Z");
       expect(result).toBe("day");
     });
   });
@@ -893,7 +893,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-15T00:00:00.000Z";
         const endDate = "2024-01-19T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -919,7 +919,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-30T00:00:00.000Z";
         const endDate = "2024-02-02T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -947,7 +947,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-15T00:00:00.000Z"; // Monday
         const endDate = "2024-01-28T23:59:59.999Z"; // Sunday
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -973,7 +973,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-29T00:00:00.000Z"; // Monday
         const endDate = "2024-02-11T23:59:59.999Z"; // Sunday
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -1001,7 +1001,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-15T00:00:00.000Z";
         const endDate = "2024-03-15T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -1030,7 +1030,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-15T00:00:00.000Z";
         const endDate = "2024-01-21T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
@@ -1063,7 +1063,7 @@ describe("EventsInsights", () => {
         const startDate = "2024-01-30T00:00:00.000Z";
         const endDate = "2024-02-05T23:59:59.999Z";
 
-        const ranges = EventsInsights.getDateRanges({
+        const ranges = getDateRanges({
           startDate,
           endDate,
           timeZone,
